refactor: remove redundant file
